refactor(*): fix export declaration
