+++ src/shared/types/index.ts
@@ -1,3 +1,6 @@
-import { ChromeMessageTypes } from './chrome-message-types';
+import { BaseChromeMessage, ChromeMessageTypes } from './chrome-message-types';
+import { SyncStorage, SyncStorageProps } from "./sync-storage";
+
 
 export { ChromeMessageTypes };
+export type { SyncStorage, BaseChromeMessage, SyncStorageProps };
