+++ src/shared/utils/safely-get-from-storage.ts
@@ -1,4 +1,4 @@
-import { type SyncStorage, type SyncStorageProps } from '../types/sync-storage';
+import { type SyncStorage, type SyncStorageProps } from '../types';
 
 /**
  * Picks up safely data from sync storage by prop.
